+++ api/src/modules/Amenity/Amenity.controller.ts
@@ -1,6 +1,8 @@
 import { NextFunction, Request, Response } from "express";
 import NotFoundError from "../../middleware/error/NotFoundError";
 import Amenity from "./Amenity.model";
+import AuthError from "../../middleware/error/AuthError";
+import { AuthRequest } from "../../middleware/auth/authMiddleware";
 
 const getAmenities = async (
 	req: Request,
@@ -38,9 +40,15 @@ const createAmenity = async (
 	next: NextFunction
 ) => {
 	try {
-		const amenity = new Amenity({ ...req.body });
-		const result = await amenity.save();
-		res.status(200).json(result);
+		const { user } = req as AuthRequest;
+
+		if (user.role === "Admin") {
+			const amenity = new Amenity({ ...req.body });
+			const result = await amenity.save();
+			res.status(200).json(result);
+		} else {
+			throw new AuthError();
+		}
 	} catch (err) {
 		next(err);
 	}
@@ -52,15 +60,21 @@ const updateAmenity = async (
 	next: NextFunction
 ) => {
 	try {
+		const { user } = req as AuthRequest;
 		const { id } = req.params;
-		const amenity = await Amenity.findOneAndUpdate({ _id: id }, req.body, {
-			new: true,
-			runValidators: true,
-		});
-		if (!amenity) {
-			throw new NotFoundError("Amenity not found");
+
+		if (user.role === "Admin") {
+			const amenity = await Amenity.findOneAndUpdate({ _id: id }, req.body, {
+				new: true,
+				runValidators: true,
+			});
+			if (!amenity) {
+				throw new NotFoundError("Amenity not found");
+			}
+			res.json(amenity);
+		} else {
+			throw new AuthError();
 		}
-		res.json(amenity);
 	} catch (err) {
 		next(err);
 	}
@@ -72,12 +86,18 @@ const deleteAmenity = async (
 	next: NextFunction
 ) => {
 	try {
+		const { user } = req as AuthRequest;
 		const { id } = req.params;
-		const amenity = await Amenity.findOneAndDelete({ _id: id });
-		if (!amenity) {
-			throw new NotFoundError("Amenity not found");
+
+		if (user.role === "Admin") {
+			const amenity = await Amenity.findOneAndDelete({ _id: id });
+			if (!amenity) {
+				throw new NotFoundError("Amenity not found");
+			}
+			res.json({});
+		} else {
+			throw new AuthError();
 		}
-		res.json({});
 	} catch (err) {
 		next(err);
 	}
